feat: move general/guest paths to guests
