+++ src/apis/guests/_id@string/index.ts
@@ -1,5 +1,5 @@
 import { Guest } from "../@types";
-import { AuthToken } from "../../../../@types";
+import { AuthToken } from "../../@types";
 
 export interface Methods {
   /**
